Reorganize directory structure.
