+++ src/recorder/condition_variable_wrapper.h
