+++ src/tracer/lightstep_immutable_span_context.cpp
@@ -1,4 +1,4 @@
-#include "lightstep_immutable_span_context.h"
+#include "tracer/lightstep_immutable_span_context.h"
 
 namespace lightstep {
 //------------------------------------------------------------------------------
