+++ src/tracer/propagation.cpp
@@ -1,4 +1,4 @@
-#include "propagation.h"
+#include "tracer/propagation.h"
 #include <lightstep/base64/base64.h>
 #include <algorithm>
 #include <array>
